cnxml/upload.zip now synced if necessary.
When a view is invoked that needs a complete upload.zip or index.cnxml,
and an outstanding index.html_ is left behind by an editing session, it is
converted, validated and the aforementioned files are updated.

In addition, fixed a bug that caused index.html to multiply in the zip file.