Fixes for static page
Add jquery, this is added in base.pt in the pyramid product. It is added again
here but with a tal instruction that will remove it if used as a tal template.

Fix paths for some aloha files. Probably a temporary fix for now.

Don't attempt to save in the static version.

Hardcode the various urls as global javascript variables, once again with
a tal instruction that places the correct value here when used as a template.