Metadata updated in index.cnxml every time
The metadata is kept as part of the editing session. This can currently only
be edited on the final /metadata page, but in theory we can move this
functionality to any place, as long as it updates the editing session with
the new metadata. Whenever index.cnxml is generated, it will include the
metadata, so that a downloaded copy will also be complete, or at least as
complete as currently possible.

Presently, only featured links are embedded in the cnxml. This needs to be
expanded to include all the metadata.